feat: allow table data to expire even if no cookie maxAge is set
Add `sessionMaxAge` property that let's you configure when the table
data is supposed to expire when no max age is configured.

By default, react-router only sets session data to expire when the
cookie has a maxAge (or expires date) set. When using session cookies,
such an expiry time does not exist, and in theory, such cookies can be
kept alive by the browser indefinitely.
However, it can still be desired to let the session data in the table
expire at some point in the near or distant future, which can be set
with this new property.